[PATRO23-7] move all business logic to HandlerSignIn
